Add contacts filter functional tests
